refactor: move fetch function to msalClient
