+++ src/hooks/useAnalogueModels.tsx
@@ -1,7 +1,6 @@
 import { useEffect, useState } from 'react'
 import { callApi } from '../api/callApi'
-import { loginRequest } from '../auth/authConfig'
-import msalInstance from '../auth/msalClient'
+import { fetchAccessToken } from '../auth/msalClient'
 import { AnalogueModel } from '../models/models'
 
 /**
@@ -12,25 +11,12 @@ export const useAnalogueModels = () => {
   const [models, setModels] = useState<AnalogueModel[]>([])
 
   useEffect(() => {
-    const request = {
-      ...loginRequest,
-      account: msalInstance.getAllAccounts()[0],
-    }
-
     function fetchAnalogueModels() {
-      msalInstance
-        .acquireTokenSilent(request)
-        .then((response) => response.accessToken)
-        .then((token) =>
-          callApi(token, '/analogueModels')
-            .then((response) => response.value)
-            .then((models) => setModels(models))
-        )
-        .catch(() =>
-          msalInstance
-            .acquireTokenPopup(request)
-            .then((response) => response.accessToken)
-        )
+      fetchAccessToken().then((token) =>
+        callApi(token, '/api/analogueModels')
+          .then((response) => response.value)
+          .then((models) => setModels(models))
+      )
     }
 
     if (models.length < 1) fetchAnalogueModels()
