+++ src/index.tsx
@@ -2,10 +2,10 @@ import { MsalProvider } from '@azure/msal-react'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import { RouterProvider } from 'react-router-dom'
-import msalInstance from './auth/msalClient'
+import { msalInstance } from './auth/msalClient'
+import './index.css'
 import reportWebVitals from './reportWebVitals'
 import { router } from './router'
-import './index.css'
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement)
 root.render(
